+++ lua/terrortown/lang/en.lua
@@ -2343,4 +2343,10 @@ Changes to this value will not take effect until players reconnect or the map ch
 L.label_roles_derandomize_min_weight = "Derandomization minimum weight"
 
 -- 2024-08-17
+L.name_button_default = "Button"
+L.name_button_rotating = "Lever"
+
+L.button_default = "Press [{usekey}] to trigger"
+L.button_rotating = "Press [{usekey}] to flip"
+
 L.undefined_key = "???"