+++ lua/ttt2/libraries/targetid.lua
@@ -198,7 +198,6 @@ function targetid.HUDDrawTargetIDTButtons(tData)
     if
         not IsValid(client)
         or not client:IsTerror()
-        or not client:Alive()
         or not IsValid(ent)
         or ent:GetClass() ~= "ttt_traitor_button"
         or tData:GetEntityDistance() > ent:GetUsableRange()
@@ -320,7 +319,6 @@ function targetid.HUDDrawTargetIDWeapons(tData)
     if
         not IsValid(client)
         or not client:IsTerror()
-        or not client:Alive()
         or not IsValid(ent)
         or tData:GetEntityDistance() > 100
         or not ent:IsWeapon()
@@ -596,6 +594,42 @@ function targetid.HUDDrawTargetIDRagdolls(tData)
 end
 
 ---
+-- This function handles looking at buttons and adds specific descriptions
+-- @param TARGET_DATA tData The object to be used in the hook
+-- @realm client
+function targetid.HUDDrawTargetIDButtons(tData)
+    local client = LocalPlayer()
+    local ent = tData:GetEntity()
+
+    if
+        not IsValid(client)
+        or not client:IsTerror()
+        or not IsValid(ent)
+        or not ent:IsButton()
+        or tData:GetEntityDistance() > 90
+    then
+        return
+    end
+
+    -- enable targetID rendering
+    tData:EnableText()
+    tData:EnableOutline()
+    tData:SetOutlineColor(client:GetRoleColor())
+
+    tData:SetKey(input.GetKeyCode(key_params.usekey))
+
+    if ent:IsDefaultButton() then
+        tData:SetTitle(TryT("name_button_default"))
+        tData:SetSubtitle(ParT("button_default", key_params))
+    end
+
+    if ent:IsRotatingButton() then
+        tData:SetTitle(TryT("name_button_rotating"))
+        tData:SetSubtitle(ParT("button_rotating", key_params))
+    end
+end
+
+---
 -- This function handles looking at doors and adds specific descriptions
 -- @param TARGET_DATA tData The object to be used in the hook
 -- @realm client
@@ -606,7 +640,6 @@ function targetid.HUDDrawTargetIDDoors(tData)
     if
         not IsValid(client)
         or not client:IsTerror()
-        or not client:Alive()
         or not IsValid(ent)
         or not ent:IsDoor()
         or not ent:PlayerCanOpenDoor()